+++ wayland-backend/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Do not use `downcast_rs` traits in API
 - Remove depreacted raw-window-handle 0.5 support
 - Split `Message`/`Argument` to have `OwnedMessage`/`OwnedArgument`
+- Use `&ObjectId`/`&ClientId` in arguments instead of `ObjectId`/`ClientId`
 
 #### Bugfixes
 - client/sys: Fix deadlock if `Backend` is used in `ObjectData::destroyed`